Add workaround for a bug in top-morph displaying
If a morph is in hand and it gets hit by a snowflake, it was damaged. This workaround respects top-morph damage for selecting hands to draw, but it is still not perfect, because interference between hands and hands, or between top-morphs and top-morphs will not be detected. A better approach would need to select and save-patch morphs at once.